# frontend/docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Orient-advertise-backend API",
    "description" : "Multi-module Spring Boot platform for TV-box content delivery — device management, content uploads + transcoding, playlists, scheduling, remote actions, incident lifecycle, proof-of-play, aggregated reporting, Excel export, and external partner integrations.\n\n**Authentication.** Most endpoints require a JWT access token (`Authorization: Bearer <token>`) obtained from `POST /auth/login`. External integration endpoints under `/service/external/**` use an `X-API-Key` header instead — see the *External* group.\n\n**Sensitive endpoints.** Operations marked with `[SENSITIVE]` in their summary affect production state, expose user data, or irreversibly delete records. Audit logging is enabled and the operator's username is recorded.\n",
    "contact" : {
      "name" : "orientadvertise.services"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "1.0.55"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Current host"
  }, {
    "url" : "https://api.example.com",
    "description" : "Production"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/users/{userId}/content/{contentFileId}" : {
      "post" : {
        "tags" : [ "Admin" ],
        "operationId" : "linkContent",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "contentFileId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "delete" : {
        "tags" : [ "Admin" ],
        "operationId" : "unlinkContent",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "contentFileId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/incidents/{id}/resolve" : {
      "post" : {
        "tags" : [ "incident-controller" ],
        "operationId" : "resolve",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IncidentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devices/{id}/heartbeat" : {
      "post" : {
        "tags" : [ "device-controller" ],
        "operationId" : "heartbeat",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/HeartbeatRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HeartbeatResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/content/{id}/retranscode" : {
      "post" : {
        "tags" : [ "content-controller" ],
        "summary" : "Re-queue transcoding for a stuck or failed content file",
        "operationId" : "retranscode",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "403" : {
            "description" : "Caller lacks ADMIN/OPERATOR",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetranscodeResult"
                }
              }
            }
          },
          "404" : {
            "description" : "Unknown content file (or soft-deleted)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetranscodeResult"
                }
              }
            }
          },
          "409" : {
            "description" : "Status is not UPLOADED/FAILED, the raw object is missing, or another worker just claimed it — see `message`",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetranscodeResult"
                }
              }
            }
          },
          "200" : {
            "description" : "Claimed and queued; returns the new status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetranscodeResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/upload" : {
      "post" : {
        "tags" : [ "content-controller" ],
        "summary" : "Upload a raw content file",
        "operationId" : "upload_1",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "urgent",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "required" : [ "file" ],
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "400" : {
            "description" : "Validation failed (file type, size, name)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller lacks ADMIN/OPERATOR",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "202" : {
            "description" : "Upload accepted; transcoding queued",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadResponse"
                },
                "examples" : {
                  "Bound to project (urgent)" : {
                    "description" : "Bound to project (urgent)",
                    "value" : {
                      "fileId" : 1042,
                      "status" : "UPLOADED",
                      "storageKey" : "raw/2026/01/abc123.mp4",
                      "urgent" : true,
                      "projectId" : 7,
                      "webSocketPush" : {
                        "sent" : 12,
                        "skipped" : 0,
                        "failed" : 1
                      },
                      "message" : "Urgent upload accepted; transcoding queued at front + WebSocket pushed"
                    }
                  },
                  "Orphan upload (no project)" : {
                    "description" : "Orphan upload (no project)",
                    "value" : {
                      "fileId" : 1043,
                      "status" : "UPLOADED",
                      "storageKey" : "raw/2026/01/def456.mp4",
                      "urgent" : false,
                      "projectId" : null,
                      "webSocketPush" : null,
                      "message" : "Upload accepted as orphan content (no project bound); attach a project via PATCH /api/content/{id}/project. Transcoding in progress"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/refresh" : {
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "Rotate the refresh cookie and mint a fresh access token",
        "description" : "Single-use refresh: the supplied cookie value is invalidated and replaced. Reuse of a previously rotated id is treated as a possible replay and invalidates the entire token family (every refresh token issued under the same login session). Role changes take effect here — the new access token is minted with the user's CURRENT role. The refresh token is read from the `refresh_token` cookie; a fresh cookie is set on the response.\n",
        "operationId" : "refresh",
        "parameters" : [ {
          "name" : "refresh_token",
          "in" : "cookie",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "401" : {
            "description" : "Cookie missing, token unknown, expired, or replay detected",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "New access token in body, rotated refresh token in cookie",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
...
    "/api/content/{id}/project" : {
      "patch" : {
        "tags" : [ "content-controller" ],
        "summary" : "Assign or clear the project on a content file",
        "operationId" : "assignProject",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AssignProjectRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "403" : {
            "description" : "Caller lacks ADMIN/OPERATOR"
          },
          "404" : {
            "description" : "Unknown content file (or already soft-deleted), or unknown project"
          },
          "204" : {
            "description" : "Project bound (or cleared if null)"
          }
        }
      }
    },
...
    "/api/stats/content/{contentFileId}" : {
      "get" : {
        "tags" : [ "stats-controller" ],
        "operationId" : "getContentStats",
        "parameters" : [ {
          "name" : "contentFileId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "deviceId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentStatsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "HeartbeatRequest" : {
        "type" : "object",
        "properties" : {
          "contentVersion" : {
            "type" : "string"
          },
          "volume" : {
            "type" : "integer",
            "format" : "int32"
          },
          "remote" : {
            "$ref" : "#/components/schemas/RemoteCapabilityDto"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "JWT access token from `POST /auth/login`. Roles encoded in the token: ADMIN, OPERATOR, VIEWER, ADVERTISER.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKeyAuth" : {
        "type" : "apiKey",
        "description" : "External-integration API key. Issued by ADMIN via `POST /service/admin/service-keys`. Rate-limited 100 req/hr per key. Revocation is immediate — no in-memory cache widens the window.",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}